+++ app/src/main/java/com/iboalali/basicrootchecker/BasicRootCheckerApplication.kt
@@ -2,6 +2,8 @@ package com.iboalali.basicrootchecker
 
 import android.app.Application
 import com.iboalali.basicrootchecker.analytics.Analytics
+import com.iboalali.basicrootchecker.billing.BillingController
+import com.iboalali.basicrootchecker.billing.createBillingController
 import com.iboalali.basicrootchecker.data.UserPreferences
 import com.iboalali.basicrootchecker.update.AppUpdateController
 import com.iboalali.basicrootchecker.update.createAppUpdateController
@@ -10,6 +12,8 @@ import com.telemetrydeck.sdk.TelemetryDeck
 class BasicRootCheckerApplication : Application() {
 
     val appUpdateController: AppUpdateController by lazy { createAppUpdateController(this) }
+
+    val billingController: BillingController by lazy { createBillingController(this) }
 
     override fun onCreate() {
         super.onCreate()